updates to the convolution layer
